refactor fixture recover_acl_rule
